+++ packages/core/src/_test/e2e/erc20/erc20.test.ts
@@ -1,26 +1,17 @@
+import net, { type AddressInfo } from "net";
 import { rmSync } from "node:fs";
-import { Ponder } from "@/Ponder.js";
 import { ALICE, BOB } from "@/_test/constants.js";
 import { setupAnvil, setupIsolatedDatabase } from "@/_test/setup.js";
 import { simulate } from "@/_test/simulate.js";
 import { onAllEventsIndexed } from "@/_test/utils.js";
-import { buildOptions } from "@/common/options.js";
+import { start } from "@/bin/commands/start.js";
 import { range } from "@/utils/range.js";
-import request from "supertest";
+import { wait } from "@/utils/wait.js";
 import { zeroAddress } from "viem";
 import { afterEach, beforeEach, describe, expect, test } from "vitest";
 
 beforeEach(setupAnvil);
 beforeEach(setupIsolatedDatabase);
-
-const gql = async (ponder: Ponder, query: string) => {
-  const response = await request(ponder.serverService.app)
-    .post("/")
-    .send({ query: `query { ${query} }` });
-  expect(response.body.errors).toBe(undefined);
-  expect(response.statusCode).toBe(200);
-  return response.body.data;
-};
 
 afterEach(() => {
   rmSync("./src/_test/e2e/erc20/.ponder", {
@@ -34,17 +25,48 @@ afterEach(() => {
   });
 });
 
-test("erc20", async (context) => {
-  const options = buildOptions({
-    cliOptions: { root: "./src/_test/e2e/erc20", config: "ponder.config.ts" },
+function getPortFree(): Promise<number> {
+  return new Promise((res) => {
+    const srv = net.createServer();
+    srv.listen(0, () => {
+      const port = (srv.address() as AddressInfo).port;
+      srv.close(() => res(port));
+    });
   });
-  const testOptions = {
-    ...options,
-    uiEnabled: false,
-    logLevel: "error",
-    telemetryDisabled: true,
-  } as const;
+}
 
+async function waitForHealthy(port: number) {
+  return new Promise((resolve, reject) => {
+    const timeout = setTimeout(() => {
+      clearInterval(interval);
+      reject(new Error("Timed out while waiting for app to become healthy."));
+    }, 5_000);
+    const interval = setInterval(async () => {
+      const response = await fetch(`http://localhost:${port}/health`);
+      if (response.status === 200) {
+        clearTimeout(timeout);
+        clearInterval(interval);
+        resolve(undefined);
+      }
+    }, 20);
+  });
+}
+
+async function postGraphql(port: number, query: string) {
+  const response = await fetch(`http://localhost:${port}/graphql`, {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify({ query: `query { ${query} }` }),
+  });
+  return response;
+}
+
+async function getMetrics(port: number) {
+  const response = await fetch(`http://localhost:${port}/metrics`);
+  return await response.text();
+}
+
+test.only("erc20", async (context) => {
   for (const _ in range(0, 3)) {
     await simulate({
       erc20Address: context.erc20.address,
@@ -52,13 +74,22 @@ test("erc20", async (context) => {
     });
   }
 
-  const ponder = new Ponder({ options: testOptions });
-  await ponder.start(context.databaseConfig);
+  const port = await getPortFree();
 
-  await onAllEventsIndexed(ponder);
+  const cleanup = await start({
+    cliOptions: {
+      root: "./src/_test/e2e/erc20",
+      config: "ponder.config.ts",
+      port,
+    },
+  });
 
-  let accounts = await gql(
-    ponder,
+  await waitForHealthy(port);
+
+  await wait(100);
+
+  let response = await postGraphql(
+    port,
     `
     accounts {
       items {
@@ -67,7 +98,12 @@ test("erc20", async (context) => {
       }
     }
     `,
-  ).then((g) => g.accounts.items);
+  );
+
+  expect(response.status).toBe(200);
+  let body = (await response.json()) as any;
+  expect(body.errors).toBe(undefined);
+  let accounts = body.data.accounts.items;
 
   expect(accounts).toHaveLength(3);
   expect(accounts[0]).toMatchObject({
@@ -88,10 +124,8 @@ test("erc20", async (context) => {
     factoryAddress: context.factory.address,
   });
 
-  await onAllEventsIndexed(ponder);
-
-  accounts = await gql(
-    ponder,
+  response = await postGraphql(
+    port,
     `
     accounts {
       items {
@@ -100,7 +134,12 @@ test("erc20", async (context) => {
       }
     }
     `,
-  ).then((g) => g.accounts.items);
+  );
+
+  expect(response.status).toBe(200);
+  body = (await response.json()) as any;
+  expect(body.errors).toBe(undefined);
+  accounts = body.data.accounts.items;
 
   expect(accounts[0]).toMatchObject({
     id: zeroAddress,
@@ -115,65 +154,65 @@ test("erc20", async (context) => {
     balance: "0",
   });
 
-  await ponder.kill();
+  await cleanup();
 });
 
-const shouldSkip = process.env.DATABASE_URL === undefined;
+// const shouldSkip = process.env.DATABASE_URL === undefined;
 
-// Fix this once it's easier to have per-command kill functions in Ponder.ts.
-describe.skipIf(shouldSkip)("postgres database", () => {
-  test.todo("ponder serve", async (context) => {
-    const options = buildOptions({
-      cliOptions: { root: "./src/_test/e2e/erc20", config: "ponder.config.ts" },
-    });
-    const testOptions = {
-      ...options,
-      uiEnabled: false,
-      logLevel: "error",
-      telemetryDisabled: true,
-    } as const;
+// // Fix this once it's easier to have per-command kill functions in Ponder.ts.
+// describe.skipIf(shouldSkip)("postgres database", () => {
+//   test.todo("ponder serve", async (context) => {
+//     const options = buildOptions({
+//       cliOptions: { root: "./src/_test/e2e/erc20", config: "ponder.config.ts" },
+//     });
+//     const testOptions = {
+//       ...options,
+//       uiEnabled: false,
+//       logLevel: "error",
+//       telemetryDisabled: true,
+//     } as const;
 
-    for (const _ in range(0, 3)) {
-      await simulate({
-        erc20Address: context.erc20.address,
-        factoryAddress: context.factory.address,
-      });
-    }
+//     for (const _ in range(0, 3)) {
+//       await simulate({
+//         erc20Address: context.erc20.address,
+//         factoryAddress: context.factory.address,
+//       });
+//     }
 
-    const ponder = new Ponder({ options: testOptions });
-    await ponder.start(context.databaseConfig);
-    await onAllEventsIndexed(ponder);
+//     const ponder = new Ponder({ options: testOptions });
+//     await ponder.start(context.databaseConfig);
+//     await onAllEventsIndexed(ponder);
 
-    const ponderServe = new Ponder({ options: testOptions });
-    await ponderServe.serve(context.databaseConfig);
+//     const ponderServe = new Ponder({ options: testOptions });
+//     await ponderServe.serve(context.databaseConfig);
 
-    const accounts = await gql(
-      ponderServe,
-      `
-      accounts {
-        items {
-          id
-          balance
-        }
-      }
-      `,
-    ).then((g) => g.accounts.items);
+//     const accounts = await gql(
+//       ponderServe,
+//       `
+//       accounts {
+//         items {
+//           id
+//           balance
+//         }
+//       }
+//       `,
+//     ).then((g) => g.accounts.items);
 
-    expect(accounts).toHaveLength(3);
-    expect(accounts[0]).toMatchObject({
-      id: zeroAddress,
-      balance: (-4 * 10 ** 18).toString(),
-    });
-    expect(accounts[1]).toMatchObject({
-      id: BOB.toLowerCase(),
-      balance: (4 * 10 ** 18).toString(),
-    });
-    expect(accounts[2]).toMatchObject({
-      id: ALICE.toLowerCase(),
-      balance: "0",
-    });
+//     expect(accounts).toHaveLength(3);
+//     expect(accounts[0]).toMatchObject({
+//       id: zeroAddress,
+//       balance: (-4 * 10 ** 18).toString(),
+//     });
+//     expect(accounts[1]).toMatchObject({
+//       id: BOB.toLowerCase(),
+//       balance: (4 * 10 ** 18).toString(),
+//     });
+//     expect(accounts[2]).toMatchObject({
+//       id: ALICE.toLowerCase(),
+//       balance: "0",
+//     });
 
-    await ponderServe.kill();
-    await ponder.kill();
-  });
-});
+//     await ponderServe.kill();
+//     await ponder.kill();
+//   });
+// });
