+++ packages/core/src/_test/e2e/erc20/ponder.config.ts
@@ -6,7 +6,19 @@ import { erc20ABI } from "../../generated.js";
 
 const poolId = Number(process.env.VITEST_POOL_ID ?? 1);
 
+function getDatabase() {
+  if (process.env.DATABASE_URL) {
+    const databaseUrl = new URL(process.env.DATABASE_URL);
+    databaseUrl.pathname = `/vitest_${poolId}`;
+    const connectionString = databaseUrl.toString();
+    return { kind: "postgres", connectionString } as const;
+  } else {
+    return { kind: "sqlite" } as const;
+  }
+}
+
 export default createConfig({
+  database: getDatabase(),
   networks: {
     mainnet: {
       chainId: 1,
